Refactor getCurrentUser function
